third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc-redirect.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<body>
<script>
    document.body.innerHTML += '<img id="redirect" src="http://127.0.0.1:8000/resources/slow-image.php?redirect_name=square100.png&mimeType=image/png&sleep=100">';
    jsTestIsAsync = true;
    var img_redirect = window.document.getElementById("redirect");
    img_redirect.onload = function() {
        shouldBe('img_redirect.currentSrc', '"http://127.0.0.1:8000/resources/slow-image.php?redirect_name=square100.png&mimeType=image/png&sleep=100"');
    };
    window.onload = function() {
        finishJSTest();
    };
</script>
</body>
